This commit message contains instructions helpful for merging.
The global defintion of ints_per_state is as follow:
ints_per_state = ((*width * *height) / 32) + ((*width * *height)%32) > 0 ? 1 : 0;

I modified the functions grouped near the top of ice.c for moving (the headers didn't change) and made a bunch of appropriate inline functions. I also modified how the the pbm's are read in the read_pbm function. I duplicated the static functions that are similar to those that exist in ice for the index and offset and mask purposes. This was because the ones in ice.c rely on a local variable that isn't defined at the time read_pbm is called, so I had to duplicate them.